Move ConsensusAlgorithm related functions to mpi_consensus_algorithm.h
