+++ include/deal.II/base/mpi_compute_index_owner_internal.h
@@ -19,6 +19,7 @@
 #include <deal.II/base/config.h>
 
 #include <deal.II/base/mpi.h>
+#include <deal.II/base/mpi_consensus_algorithm.h>
 
 DEAL_II_NAMESPACE_OPEN
